Move completion provider to standalone file
